+++ src/views/Sale/components/HeaderControl/index.tsx
@@ -19,7 +19,7 @@ import { useWindowSize } from 'src/hooks/useWindowSize'
 import { Sale } from 'src/interfaces/Sale'
 
 // Aqua Utils
-import { formatBigInt } from 'src/utils/Defaults'
+import { formatBigInt } from 'src/utils/helpers'
 
 // hooks
 import { BigNumber } from '@ethersproject/bignumber'
