+++ src/views/Sale/components/PlaceBidForm/index.spec.tsx
@@ -19,7 +19,7 @@ import { theme } from 'src/styles/theme'
 import { PlaceBidForm } from 'src/views/Sale/components/PlaceBidForm/index'
 import { calculateClearingPrice } from 'src/aqua/price'
 //defaults
-import { getSaleDefault } from 'src/utils/Defaults'
+import { getSaleDefault } from 'src/utils/helpers'
 
 // Extends dayjs
 dayjs.extend(DayjsRelativeTime)
