+++ src/views/Sale/components/TokenFooter/index.spec.tsx
@@ -7,7 +7,7 @@ import '@testing-library/jest-dom/extend-expect'
 import { TokenFooter } from 'src/views/Sale/components/TokenFooter/index'
 
 // default
-import { getSaleDefault, getSaleDetailsDefault } from 'src/utils/Defaults'
+import { getSaleDefault, getSaleDetailsDefault } from 'src/utils/helpers'
 
 describe('TokenFooter', () => {
   test('should display multiple texts on Footer component', () => {
